#150: Move endpoint tests to their own files/classes
