+++ test/icat/test_standard_endpoints.py
@@ -1,9 +1,6 @@
-import pytest
-
 from datagateway_api.src.main import api
 from datagateway_api.src.main import app
 from datagateway_api.src.resources.entities.entity_map import endpoints
-from test.icat.test_query import prepare_icat_data_for_assertion
 from test.test_base import FlaskAppTest
 
 
@@ -31,444 +28,3 @@ class TestStandardEndpoints:
             assert findone_endpoint_exist
         """
         pass
-
-    def test_valid_get_with_filters(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations?where={"title": {"eq": "Test data for the Python ICAT'
-            ' Backend on DataGateway API"}}',
-            headers=valid_credentials_header,
-        )
-        response_json = prepare_icat_data_for_assertion(test_response.json)
-
-        assert response_json == single_investigation_test_data
-
-    def test_valid_no_results_get_with_filters(
-        self, flask_test_app, valid_credentials_header,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations?where={"title": {"eq": "This filter should cause a 404 for'
-            'testing purposes..."}}',
-            headers=valid_credentials_header,
-        )
-
-        assert test_response.status_code == 404
-
-    @pytest.mark.usefixtures("multiple_investigation_test_data")
-    def test_valid_get_with_filters_distinct(
-        self, flask_test_app, valid_credentials_header,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations?where={"title": {"like": "Test data for the Python ICAT'
-            ' Backend on DataGateway API"}}&distinct="title"',
-            headers=valid_credentials_header,
-        )
-
-        expected = [
-            {
-                "title": f"Test data for the Python ICAT Backend on DataGateway API {i}"
-                for i in range(5)
-            },
-        ]
-
-        for title in expected:
-            assert title in test_response.json
-
-    def test_limit_skip_merge_get_with_filters(
-        self,
-        flask_test_app,
-        valid_credentials_header,
-        multiple_investigation_test_data,
-    ):
-        skip_value = 1
-        limit_value = 2
-
-        test_response = flask_test_app.get(
-            '/investigations?where={"title": {"like": "Test data for the Python ICAT'
-            ' Backend on DataGateway API"}}'
-            f'&skip={skip_value}&limit={limit_value}&order="id ASC"',
-            headers=valid_credentials_header,
-        )
-        response_json = prepare_icat_data_for_assertion(test_response.json)
-
-        filtered_investigation_data = []
-        filter_count = 0
-        while filter_count < limit_value:
-            filtered_investigation_data.append(
-                multiple_investigation_test_data.pop(skip_value),
-            )
-            filter_count += 1
-
-        assert response_json == filtered_investigation_data
-
-    def test_valid_create_data(self, flask_test_app, valid_credentials_header):
-        create_investigations_json = [
-            {
-                "name": "Test Data for API Testing, Data Creation 1",
-                "title": "Test data for the Python ICAT Backend on DataGateway API",
-                "summary": "Test data for DataGateway API testing",
-                "releaseDate": "2020-03-03 08:00:08",
-                "startDate": "2020-02-02 09:00:09",
-                "endDate": "2020-02-03 10:00:10",
-                "visitId": "Data Creation Visit",
-                "doi": "DataGateway API Test DOI",
-                "facility": 1,
-                "type": 1,
-            },
-            {
-                "name": "Test Data for API Testing, Data Creation 2",
-                "title": "Test data for the Python ICAT Backend on DataGateway API",
-                "summary": "Test data for DataGateway API testing",
-                "releaseDate": "2020-03-03 08:00:08",
-                "startDate": "2020-02-02 09:00:09",
-                "endDate": "2020-02-03 10:00:10",
-                "visitId": "Data Creation Visit",
-                "doi": "DataGateway API Test DOI",
-                "facility": 1,
-                "type": 1,
-            },
-        ]
-
-        test_response = flask_test_app.post(
-            "/investigations",
-            headers=valid_credentials_header,
-            json=create_investigations_json,
-        )
-
-        test_data_ids = []
-        for investigation_request, investigation_response in zip(
-            create_investigations_json, test_response.json,
-        ):
-            investigation_request.pop("facility")
-            investigation_request.pop("type")
-            test_data_ids.append(investigation_response["id"])
-
-        response_json = prepare_icat_data_for_assertion(
-            test_response.json, remove_id=True,
-        )
-
-        assert create_investigations_json == response_json
-
-        # Delete the entities created by this test
-        for investigation_id in test_data_ids:
-            flask_test_app.delete(
-                f"/investigations/{investigation_id}", headers=valid_credentials_header,
-            )
-
-    def test_valid_boundary_create_data(self, flask_test_app, valid_credentials_header):
-        """Create a single investigation, as opposed to multiple"""
-
-        create_investigation_json = {
-            "name": "Test Data for API Testing, Data Creation 0",
-            "title": "Test data for the Python ICAT Backend on the API",
-            "summary": "Test data for DataGateway API testing",
-            "releaseDate": "2020-03-03 08:00:08",
-            "startDate": "2020-02-02 09:00:09",
-            "endDate": "2020-02-03 10:00:10",
-            "visitId": "Data Creation Visit",
-            "doi": "DataGateway API Test DOI",
-            "facility": 1,
-            "type": 1,
-        }
-
-        test_response = flask_test_app.post(
-            "/investigations",
-            headers=valid_credentials_header,
-            json=create_investigation_json,
-        )
-
-        create_investigation_json.pop("facility")
-        create_investigation_json.pop("type")
-        created_test_data_id = test_response.json[0]["id"]
-
-        response_json = prepare_icat_data_for_assertion(
-            test_response.json, remove_id=True,
-        )
-
-        assert [create_investigation_json] == response_json
-
-        flask_test_app.delete(
-            f"/investigations/{created_test_data_id}", headers=valid_credentials_header,
-        )
-
-    def test_invalid_create_data(self, flask_test_app, valid_credentials_header):
-        """An investigation requires a minimum of: name, visitId, facility, type"""
-
-        invalid_request_body = {
-            "title": "Test Title for DataGateway API Backend testing",
-        }
-
-        test_response = flask_test_app.post(
-            "/investigations",
-            headers=valid_credentials_header,
-            json=invalid_request_body,
-        )
-
-        assert test_response.status_code == 400
-
-    def test_invalid_existing_data_create(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        """This test targets raising ICATObjectExistsError, causing a 400"""
-
-        # entity.as_dict() removes details about facility and type, hence they're
-        # hardcoded here instead of using sinle_investigation_test_data
-        existing_object_json = {
-            "name": single_investigation_test_data[0]["name"],
-            "title": single_investigation_test_data[0]["title"],
-            "visitId": single_investigation_test_data[0]["visitId"],
-            "facility": 1,
-            "type": 1,
-        }
-
-        test_response = flask_test_app.post(
-            "/investigations",
-            headers=valid_credentials_header,
-            json=existing_object_json,
-        )
-
-        assert test_response.status_code == 400
-
-    def test_valid_multiple_update_data(
-        self,
-        flask_test_app,
-        valid_credentials_header,
-        multiple_investigation_test_data,
-    ):
-        expected_doi = "Test Data Identifier"
-        expected_summary = "Test Summary"
-
-        update_data_list = []
-
-        for investigation in multiple_investigation_test_data:
-            investigation["doi"] = expected_doi
-            investigation["summary"] = expected_summary
-
-            update_entity = {
-                "id": investigation["id"],
-                "doi": expected_doi,
-                "summary": expected_summary,
-            }
-            update_data_list.append(update_entity)
-
-        test_response = flask_test_app.patch(
-            "/investigations", headers=valid_credentials_header, json=update_data_list,
-        )
-        response_json = prepare_icat_data_for_assertion(test_response.json)
-
-        assert response_json == multiple_investigation_test_data
-
-    def test_valid_boundary_update_data(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        """ Request body is a dictionary, not a list of dictionaries"""
-
-        expected_doi = "Test Data Identifier"
-        expected_summary = "Test Summary"
-
-        update_data_json = {
-            "id": single_investigation_test_data[0]["id"],
-            "doi": expected_doi,
-            "summary": expected_summary,
-        }
-        single_investigation_test_data[0]["doi"] = expected_doi
-        single_investigation_test_data[0]["summary"] = expected_summary
-
-        test_response = flask_test_app.patch(
-            "/investigations", headers=valid_credentials_header, json=update_data_json,
-        )
-        response_json = prepare_icat_data_for_assertion(test_response.json)
-
-        assert response_json == single_investigation_test_data
-
-    def test_invalid_missing_update_data(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        """There should be an ID in the request body to know which entity to update"""
-
-        update_data_json = {
-            "doi": "Test Data Identifier",
-            "summary": "Test Summary",
-        }
-
-        test_response = flask_test_app.patch(
-            "/investigations", headers=valid_credentials_header, json=update_data_json,
-        )
-
-        assert test_response.status_code == 400
-
-    @pytest.mark.parametrize(
-        "update_key, update_value",
-        [
-            pytest.param("invalidAttr", "Some Value", id="invalid attribute"),
-            pytest.param("modId", "simple/root", id="meta attribute update"),
-        ],
-    )
-    def test_invalid_attribute_update(
-        self,
-        flask_test_app,
-        valid_credentials_header,
-        single_investigation_test_data,
-        update_key,
-        update_value,
-    ):
-        invalid_update_data_json = {
-            "id": single_investigation_test_data[0]["id"],
-            update_key: update_value,
-        }
-
-        test_response = flask_test_app.patch(
-            "/investigations",
-            headers=valid_credentials_header,
-            json=invalid_update_data_json,
-        )
-
-        assert test_response.status_code == 400
-
-    def test_valid_findone_with_filters(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations/findone?where={"title": {"eq": "Test data for the Python'
-            ' ICAT Backend on DataGateway API"}}',
-            headers=valid_credentials_header,
-        )
-        response_json = prepare_icat_data_for_assertion([test_response.json])
-
-        assert response_json == single_investigation_test_data
-
-    def test_valid_no_results_findone_with_filters(
-        self, flask_test_app, valid_credentials_header,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations/findone?where={"title": {"eq": "This filter should cause a'
-            '404 for testing purposes..."}}',
-            headers=valid_credentials_header,
-        )
-
-        assert test_response.status_code == 404
-
-    @pytest.mark.usefixtures("single_investigation_test_data")
-    def test_valid_count_with_filters(self, flask_test_app, valid_credentials_header):
-        test_response = flask_test_app.get(
-            '/investigations/count?where={"title": {"eq": "Test data for the Python'
-            ' ICAT Backend on DataGateway API"}}',
-            headers=valid_credentials_header,
-        )
-
-        assert test_response.json == 1
-
-    def test_valid_no_results_count_with_filters(
-        self, flask_test_app, valid_credentials_header,
-    ):
-        test_response = flask_test_app.get(
-            '/investigations/count?where={"title": {"eq": "This filter should cause a'
-            '404 for testing purposes..."}}',
-            headers=valid_credentials_header,
-        )
-
-        assert test_response.json == 0
-
-    def test_valid_get_with_id(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        # Need to identify the ID given to the test data
-        investigation_data = flask_test_app.get(
-            '/investigations?where={"title": {"eq": "Test data for the Python ICAT'
-            ' Backend on DataGateway API"}}',
-            headers=valid_credentials_header,
-        )
-        test_data_id = investigation_data.json[0]["id"]
-
-        test_response = flask_test_app.get(
-            f"/investigations/{test_data_id}", headers=valid_credentials_header,
-        )
-        # Get with ID gives a dictionary response (only ever one result from that kind
-        # of request), so list around json is required for the call
-        response_json = prepare_icat_data_for_assertion([test_response.json])
-
-        assert response_json == single_investigation_test_data
-
-    def test_invalid_get_with_id(self, flask_test_app, valid_credentials_header):
-        """Request with a non-existent ID"""
-
-        final_investigation_result = flask_test_app.get(
-            '/investigations/findone?order="id DESC"', headers=valid_credentials_header,
-        )
-        test_data_id = final_investigation_result.json["id"]
-
-        # Adding 100 onto the ID to the most recent result should ensure a 404
-        test_response = flask_test_app.get(
-            f"/investigations/{test_data_id + 100}", headers=valid_credentials_header,
-        )
-
-        assert test_response.status_code == 404
-
-    def test_valid_delete_with_id(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        test_response = flask_test_app.delete(
-            f'/investigations/{single_investigation_test_data[0]["id"]}',
-            headers=valid_credentials_header,
-        )
-
-        assert test_response.status_code == 204
-
-    def test_invalid_delete_with_id(self, flask_test_app, valid_credentials_header):
-        """Request with a non-existent ID"""
-
-        final_investigation_result = flask_test_app.get(
-            '/investigations/findone?order="id DESC"', headers=valid_credentials_header,
-        )
-        test_data_id = final_investigation_result.json["id"]
-
-        # Adding 100 onto the ID to the most recent result should ensure a 404
-        test_response = flask_test_app.delete(
-            f"/investigations/{test_data_id + 100}", headers=valid_credentials_header,
-        )
-
-        assert test_response.status_code == 404
-
-    def test_valid_update_with_id(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        expected_doi = "Test Data Identifier"
-        expected_summary = "Test Summary"
-
-        update_data_json = {
-            "doi": expected_doi,
-            "summary": expected_summary,
-        }
-        single_investigation_test_data[0]["doi"] = expected_doi
-        single_investigation_test_data[0]["summary"] = expected_summary
-
-        test_response = flask_test_app.patch(
-            f"/investigations/{single_investigation_test_data[0]['id']}",
-            headers=valid_credentials_header,
-            json=update_data_json,
-        )
-        response_json = prepare_icat_data_for_assertion([test_response.json])
-
-        assert response_json == single_investigation_test_data
-
-    def test_invalid_update_with_id(
-        self, flask_test_app, valid_credentials_header, single_investigation_test_data,
-    ):
-        """This test will attempt to put `icatdb` into an invalid state"""
-
-        # DOI cannot be over 255 characters, which this string is
-        invalid_update_json = {
-            "doi": "__________________________________________________________________"
-            "_________________________________________________________________________"
-            "_________________________________________________________________________"
-            "_________________________________________________________________________",
-        }
-
-        test_response = flask_test_app.patch(
-            f"/investigations/{single_investigation_test_data[0]['id']}",
-            headers=valid_credentials_header,
-            json=invalid_update_json,
-        )
-
-        assert test_response.status_code == 400
